+++ WorldCities/WorldCitiesAPI/Controllers/CountriesController.cs
@@ -27,7 +27,7 @@ public class CountriesController : ControllerBase
   /// </summary>
   /// <returns>Return a list of countries</returns>
   [HttpGet]
-  public async Task<ActionResult<ApiResult<Country>>> GetCountries(
+  public async Task<ActionResult<ApiResult<CountryDTO>>> GetCountries(
     int pageIndex,
     int pageSize = 10,
     string? sortColumn = null,
@@ -35,8 +35,15 @@ public class CountriesController : ControllerBase
     string? filterColumn = null,
     string? filterQuery = null)
   {
-    return await ApiResult<Country>.CreateAsync(
-      _context.Countries.AsNoTracking(),
+    return await ApiResult<CountryDTO>.CreateAsync(
+      _context.Countries.AsNoTracking().Select(c => new CountryDTO()
+      {
+        Id = c.Id,
+        Name = c.Name,
+        ISO2 = c.ISO2,
+        ISO3 = c.ISO3,
+        TotCities = c.Cities!.Count()
+      }),
       pageIndex,
       pageSize,
       sortColumn,